requirements.txt
matplotlib
paho-mqtt
twilio
flask
